Added free surface model implemented with rectangular constant strength source panels. A number of new inputs were added (see Doc/Input_Description for details. A few new outputs were added (see Doc/Output_Description for details). Briefly, the FSFlag input in the config namelist activates the free surface. Set the submergence depth using the dFS parameter in the geometry namelist.
Currently, while one can use both the free surface and ground plane models simultaneously, the panel systems are solved 
independently, one after the other. This is likely to be OK when there is considerable separation between the two surfaces,
but may produce instability when the surfaces are in close proximity. This functionality hasn't been tested. In future versions,
these panel systems will likely be solved simultaneously to eliminate the potential for instability...