Modified 'HAWT' functionality of CreateTurbine
Modified the 'HAWT' functionality. Allowed for the radial distribution of vortex
lines on the blade surface to be specified directly through arguments instead of
global variables. Modified so that the airfoil ID distribution is written directly
to the .GEOM file (see iSect variable).

Note that the Test/TestGeom/TestHAWT.m is currently broken (and was broken prior to
these modifications).