Added default parameters for wake data grids.
Added default parameters for wake data grids. Previously, default parameters
only existed for wake data on a cross-sectional grid (corresponding to WakeOutFlag = 4).
This would result in a silent error when using WakeOutFlag = 2 or 3 without specifying
parameters for the wake data planes.

It may be useful in the future to add a warning/message to the user that they can specify
the paramaters of the wake data grids.

Also cleaned up some spacing in the input.f95 file.