Wake plane output now outputs grid
This updates the wake plane output functions, which suffered from a
number of problems. First, nowhere in the wake plane output file was the
location of the slice given. Second, only one wake plane was able to be
output for any given run.

This edit enables the wake to be output on a 3-D Cartesian grid with
specified dimensions and numbers of grid points in each direction.

This greatly simplifies the wake output functions at the expense of
slightly complicating the input options.