Fixed output file header linebreak bug
Fixed a bug that was causing the header for long headerfiles to be
output with a line break. This bug was caused by using list-directed
output (command that looked like write(FID,*)) in the csvwrite sub-
routine.  This was fixed by changing the output format string to '(A)'.

It should be noted that only some compilers break list-directed output
into multiple lines (Intel Fortran does at col 80, gfortran doesn't).
The fix makes the behavior more predictable across compilers.

For details on list-directed vs formatted output, see:
http://computer-programming-forum.com/49-fortran/c1fd8d5e00a7fba5.htm