Changed intervals for wake/wakegrid output
Changed the intervals that the wake and wakegrid are output. Previous
versions had some inconsistencies that would result in the wake and
wakegrid data being output at different times, even though the output
intervals were the same.

Changed the timestep indexing in the input file -- now users should
specify a start timestep of '1' instead of '0' in order to start output
at the first timestep (the first timestep in CACTUS is '1', not '0').

Updated the documentation to reflect these changes.