Improved performance of WriteWakeGridData
Significantly improved the performance of WriteWakeGridData by
rearranging the loop structure.

Previously, the induced velocities were computed and written in the
same loop, using temporary variables to store the induced velocity at
each grid point. This was extremely slow because it required the data
to be written to file before the temporary variables could be
overwritten.

The new method separates the loops so that all induced velocities are
saved to memory in one loop and are written to file in a subsequent
loop. This takes advantage of the Fortran I/O buffer to improve speed.