Fixed wake element output bug
Fixed a bug in WriteWakeData that was causing missing elements in the wake
output data. The error was caused by a problem in the indexing. The data
for the nodes is stored in X,Y,Z and U,V,W which is indexed by the
timestep number the node was generated on and the wake element index.
Prior to the bugfix, the wake element index was only being looped
through to the number of ELEMENTS, but there are more segments than
elements (num segments = num elements + num blades).

Also modified the column headers to be more informative.