Parallelized computation of grid velocity data
Parallelized the computation of grid velocity data (and of free-stream
data on grid nodes) using OpenMP. The OpenMP loop is on the y-loop,
and will be inefficiently parallelized when the number of grid points
in the y-direction is small.

In the future it may be desirable to change the parallelization loop
depending on the dimensions nx,ny,nz of the grid.