Fixed bug in wake grid output
Fixed a bug in the wake grid output that was causing garbage data to be
output when WriteWakeGridStartTimestep > 0. The problem was caused by
the placement of the Cartesian grid initialization inside an
if-statement in the WriteWakeGrid.f95 subroutine that was only executed
on NT==1.

The grid initialization was moved to the wakegrid_cns() constructor in
the wakedata module, which is called by input.f95 on program execution.

Also removed some extraneous variable declarations in
WriteWakeGridData.f95.