Fixed write problem w/ _ParamData and _RevData
Fixed a problem where the _ParamData and _RevData files were only being
written at the end of a simulation. (This was problematic when
simulations weren't being executed to completion).

The solution uses CLOSE() to close the _ParamData file and the _RevData
files before the end of the program, which forces the files to be
written from the buffer immediately.