Added wallgeom module
Added `wallgeom` module, which offloads some of the duties of the old
`wallsoln`. The `wallgeom` module is specifically for managing the
geometry of the walls. It includes functions (using the plot3d module)
to read in a multi-block Plot3D mesh as a series of `Wall` types.

At this commit, it is redundant with certain parts of `wallsoln`