Changed ground plane to use new wallsystem/quadsourcepanel
Changed the ground plane implementation to use the new wallsystem and
quadsourcepanel implementation. The mesh generation from the original
ground plane has been kept intact, however this ground plane is loaded
 into a type(WallType).

The WSolnSetup() routine, which was responsible for generating the
influence matrix and inverting it, now calls the appropriate
subroutines from the wallsystem module.