Changed index ordering in WGeomSetup()
Changed the index ordering in WGeomSetup() in order to match the index
convention for the quadsourcepanel and wallsystem modules. Each wall
wall should be indexed so that the normal vector W3Vec = W1Vec x W2Vec
points inward towards the domain. The same is true for each individual
panel on the wall.